EDM-2304: Consolidate into single standalone cmd with render subcommand
