+++ deploy/podman/flightctl-db-migrate/flightctl-db-migrate.container
@@ -23,7 +23,7 @@ Restart=on-failure
 RestartSec=30
 TimeoutStartSec=10m
 
-ExecStartPre=/usr/bin/flightctl-ground-crew render --input-file /usr/share/flightctl/flightctl-db-migrate/config.yaml.template --output-file /etc/flightctl/flightctl-db-migrate/config.yaml
+ExecStartPre=/usr/bin/flightctl-standalone render template --input-file /usr/share/flightctl/flightctl-db-migrate/config.yaml.template --output-file /etc/flightctl/flightctl-db-migrate/config.yaml
 
 [Install]
 WantedBy=flightctl.target
