+++ deploy/podman/flightctl-pam-issuer/flightctl-pam-issuer.container
@@ -17,7 +17,7 @@ Volume=/etc/flightctl/flightctl-pam-issuer/config.yaml:/root/.flightctl/config.y
 Restart=always
 RestartSec=30
 
-ExecStartPre=/usr/bin/flightctl-ground-crew render --input-file /usr/share/flightctl/flightctl-pam-issuer/config.yaml.template --output-file /etc/flightctl/flightctl-pam-issuer/config.yaml
+ExecStartPre=/usr/bin/flightctl-standalone render template --input-file /usr/share/flightctl/flightctl-pam-issuer/config.yaml.template --output-file /etc/flightctl/flightctl-pam-issuer/config.yaml
 
 [Install]
 WantedBy=flightctl.target
