+++ deploy/podman/flightctl-periodic/flightctl-periodic.container
@@ -20,7 +20,7 @@ Volume=/etc/flightctl/ssh:/etc/flightctl/ssh:ro,z
 Restart=always
 RestartSec=30
 
-ExecStartPre=/usr/bin/flightctl-ground-crew render --input-file /usr/share/flightctl/flightctl-periodic/config.yaml.template --output-file /etc/flightctl/flightctl-periodic/config.yaml
+ExecStartPre=/usr/bin/flightctl-standalone render template --input-file /usr/share/flightctl/flightctl-periodic/config.yaml.template --output-file /etc/flightctl/flightctl-periodic/config.yaml
 
 [Install]
 WantedBy=flightctl.target
