+++ deploy/scripts/deploy_quadlets.sh
@@ -9,8 +9,8 @@ source "${SCRIPT_DIR}"/shared.sh
 echo "Starting Deployment"
 
 # Render quadlet files
-make build-render-quadlets
-bin/render-quadlets --config deploy/podman/local-images.yaml
+make build-standalone
+bin/flightctl-standalone render quadlets --config deploy/podman/local-images.yaml
 
 echo "Ensuring secrets are available..."
 # Always ensure secrets exist before starting services
